+++ tests/retain-additional-fields.spec.ts
@@ -1,7 +1,5 @@
 import {test, expect} from '@playwright/test';
 import { MainPO } from './po/main-po';
-import path from 'path';
-import fs from 'node:fs/promises';
 import {PWUtils} from "./pw-utils";
 
 test.describe('retain-additional-fields', async () => {
@@ -15,7 +13,11 @@ test.describe('retain-additional-fields', async () => {
   });
 
   test('should retain contained field', async ({page}) => {
-    const {fileJson, fbJson} = await PWUtils.loadFile(page, '../cypress/fixtures/contained-example.json');
-    expect(fbJson.contained).toEqual(fileJson.contained);
+    await page.getByLabel('Start from scratch').click();
+    await page.getByRole('button', {name: 'Continue'}).click();
+    const fileJson = await PWUtils.uploadFile(page, '../cypress/fixtures/contained-example.json');
+    await expect(page.locator('input#title')).toHaveValue('Contained example');
+    const json = await PWUtils.getQuestionnaireJSON(page, 'R4');
+    expect(json.contained).toEqual(fileJson.contained);
   });
 });