+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [9.6.8] 2024-10-17
+## [9.6.9] 2024-10-17
 ### Changed
 - Added Group Item Control options.
+
+## [9.6.8] 2024-09-18
+### Fixed
+- Fixed help item showing as a node in the tree.
+### Added
+- Added support for extensions of rendering styles and rendering xhtml to
+- question text, prefix and help text fields.
 
 ## [9.6.7] 2024-10-03
 ### Changed